LPS-175543 Extract the AssetSearcher from portal-impl to the portlet service module as an internal class
